FlinkkafkaSource 配置kerberos认证 已成功 flink table kafka connector 配置kerberos认证 暂未成功

+++ src/main/java/com/zetyun/hqbank/FlinkTable2IceTable.java
@@ -1,46 +1,64 @@
 package com.zetyun.hqbank;
 
 import com.zetyun.hqbank.service.oracle.OracleService;
-import com.zetyun.hqbank.util.FileUtil;
 import com.zetyun.hqbank.util.YamlUtil;
+import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.flink.configuration.Configuration;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.table.api.EnvironmentSettings;
 import org.apache.flink.table.api.bridge.java.StreamTableEnvironment;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
-import static com.zetyun.hqbank.DDS2FlinkCDC.whiteList;
 
 public class FlinkTable2IceTable {
     private static Logger logger = LoggerFactory.getLogger(FlinkTable2IceTable.class);
-    private static final HashMap<String, HashMap<String, String>> tableNameMap = new HashMap<>();
-    public static final List<String> whiteList = Arrays.asList(new String[]{"DDS_T_ZHJ2"});
-    public static final String PATH  = "/data/rdx/application.yaml";
-    public static void main(String[] args) {
+    public static final String PATH = "D:/conf/application.yaml";
 
+    public static void main(String[] args) {
+        // kerberos 认证配置
+        String jaasConf = YamlUtil.getValueByKey(PATH, "kerberos", "jaasConf");
+        String krb5Conf = YamlUtil.getValueByKey(PATH, "kerberos", "krb5Conf");
+        String krb5Keytab = YamlUtil.getValueByKey(PATH, "kerberos", "krb5Keytab");
+        String principal = YamlUtil.getValueByKey(PATH, "kerberos", "principal");
+
+        // catalog 配置信息
         String hiveUri = YamlUtil.getValueByKey(PATH, "hadoop", "hiveUri");
         String warehouse = YamlUtil.getValueByKey(PATH, "hadoop", "warehouse");
         String hiveConfDir = YamlUtil.getValueByKey(PATH, "hadoop", "hiveConfDir");
         String hadoopConfDir = YamlUtil.getValueByKey(PATH, "hadoop", "hadoopConfDir");
+        // 读取建表语句
+        String databaseName = YamlUtil.getValueByKey(PATH, "table", "database");
+        List<String> owners = YamlUtil.getListByKey(PATH, "table", "owner");
+        // kafka 配置
+        String bootstrap = YamlUtil.getValueByKey(PATH, "kafka", "bootstrap");
+        // 白名单，如果不为空，则建立所有表的流
+        List<String> whiteList = YamlUtil.getListByKey(PATH, "table", "whiteListB");
+        String catalogName = YamlUtil.getValueByKey(PATH, "catalog", "iceberg");
+        Boolean deleteOldFlinkTable = YamlUtil.getBooleanValueByKey(PATH, "flink", "deleteOldTable");
 
-        // 读取配置文件
 //        Configuration conf = new Configuration();
 //        conf.setInteger(RestOptions.PORT, 10000);
+        // flink 指定 jaas 必须此配置 用于认证
+        System.setProperty("java.security.auth.login.config", jaasConf);
 
-        StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment();
-        env.enableCheckpointing(9000L);
+        Properties flinkProps = new Properties();
+        flinkProps.setProperty("security.kerberos.krb5-conf.path", krb5Conf);
+        flinkProps.setProperty("security.kerberos.login.keytab", krb5Keytab);
+        flinkProps.setProperty("security.kerberos.login.principal", principal);
+        flinkProps.setProperty("security.kerberos.login.contexts", "Client,KafkaClient");
+        flinkProps.setProperty("state.backend", "hashmap");
+
+        Configuration flinkConfig = new Configuration();
+        flinkConfig.addAllToProperties(flinkProps);
+        StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment(flinkConfig);
+        env.enableCheckpointing(8000L);
 
         EnvironmentSettings settings = EnvironmentSettings.newInstance().inStreamingMode().useBlinkPlanner().build();
         StreamTableEnvironment streamTableEnv = StreamTableEnvironment.create(env, settings);
-        String catalogName = "iceberg_catalog";
-        String databaseName = YamlUtil.getValueByKey(PATH, "table", "database");
-        List<String> owners = YamlUtil.getListByKey(PATH, "table", "owner");
-        String bootstrap = YamlUtil.getValueByKey(PATH, "kafka", "bootstrap");
 
         HashMap<String, HashMap<String, String>> sqlMap = null;
         OracleService oracleTrigger = new OracleService();
@@ -75,13 +93,19 @@ public class FlinkTable2IceTable {
             String kafkaSql = value.get("KAFKA_" + tableName);
             String iceSql = value.get("ICE_" + tableName);
             logger.info("kafkaSql:{},iceSql:{},tableName:{}", kafkaSql, iceSql, tableName);
-//            if (!whiteList.contains(tableName)) {// todo delete line
-//                continue;
-//            }
+
+            if (CollectionUtils.isNotEmpty(whiteList)) {
+                if (!whiteList.contains(tableName)) {
+                    continue;
+                }
+            }
+
             String sinkTable = catalogName + "." + databaseName + ".ICE_" + tableName;
             String sourceTable = databaseName + ".KAFKA_" + tableName;
-            streamTableEnv.executeSql("drop table if exists " + sourceTable);
-            streamTableEnv.executeSql("drop table if exists " + sinkTable);
+            if (deleteOldFlinkTable) {
+                streamTableEnv.executeSql("drop table if exists " + sourceTable);
+                streamTableEnv.executeSql("drop table if exists " + sinkTable);
+            }
 
             // create flink table with kafka topic
             logger.info("create flink table with kafka connector:{}", kafkaSql);
@@ -95,9 +119,6 @@ public class FlinkTable2IceTable {
             String insertSql = "insert into " + sinkTable + "  select * from " + sourceTable;
             logger.info("insert :{}", insertSql);
             streamTableEnv.executeSql(insertSql);
-
         }
-
-
     }
 }
